requirements.txt
streamlit
numpy
matplotlib
Pillow
networkx
scipy
